Standardize subscripts

collect_subscripts() standardizes subscripts to make it easier to develop [ methods, especially for data frame and array subsetting. It helps with a number of things, including:

Counting the number of subscript arguments supplied.

Transforming implicit missing subscripts into explicit NULL values.

Optional support for capturing ....

Optional support for standardizing the data frame column subsetting behavior of x[i] into x[,j].

collect_subscripts() can only be called from inside of a function. Calling collect_subscripts() in the console directly will result in an error.

The expected way to use collect_subscripts() is as the first line of a [ method. You'll collect the processed subscripts, and then use them to slice the input accordingly.

collect_subscripts(
  i,
  j,
  ...,
  column_transform = TRUE,
  env = parent.frame(),
  fn = sys.function(sys.parent())
)

Arguments

i, j
[vector]

Required vectors of subscripts.

...
[dots]

Optional additional subscripts for subsetting into higher dimensions. Names are preserved. Implicit missing subscripts are transformed into explicit NULL values.

If no ... are provided, the return value will contain an empty list in the $dots slot.

column_transform
[logical(1)]

Should a single i argument represent column subsetting? For example, this standardizes x[i] to x[,j]. This is useful when creating data frame [ methods, but array subsetting likely won't use this feature.

env
[environment]

Generally, the calling environment of the [ method. Used to detect the number of arguments supplied when the user invoked the [ method.

fn
[function]

Generally, the function corresponding to the [ method. Used to check that the [ method is valid, and to count the number of arguments alongside env.

Value

A named list of the following 4 elements containing information about the standardized subscripts.

i: The standardized i subscript. If NULL, i is considered to be missing.

j: The standardized j subscript. If NULL, j is considered to be missing.

dots: A list of the standardized ... subscripts. If no ... are supplied, an empty list is returned. If ... are supplied and any subscripts are missing, NULL values are returned in those locations. For example, x[i, j, , k] would return a list of list(NULL, k).

transformed: A logical of size 1 indicating whether the transformation of x[i] to x[,j] was performed or not.

Details

If NULL is provided as a subscript, it is standardized to integer() to prevent confusion with NULL values in the resulting info list. A NULL subscript should have the same meaning as integer().

Examples

# Generally you would replace `slicer()` with your `[.myclass` method,
# which you would then call as `x[i]`. This just collects the subscript info
# and returns a compact view of it. Normally you would use the processed
# results to slice `x`.

slicer <- function(x, i, j, ..., drop = FALSE) {
  info <- collect_subscripts(i, j, ...)
  str(info)
}

x <- "a"

# Standardize an easy case
slicer(x, 1:5, 2)

#> List of 4
#>  $ i          : int [1:5] 1 2 3 4 5
#>  $ j          : num 2
#>  $ dots       : list()
#>  $ transformed: logi FALSE

# By default, supplying just `i` standardizes `x[i]` to `x[,j]`
slicer(x, 1)

#> List of 4
#>  $ i          : NULL
#>  $ j          : num 1
#>  $ dots       : list()
#>  $ transformed: logi TRUE

# With just `j`, the implicit missing `i` value is returned as a `NULL`.
slicer(x, , 1)

#> List of 4
#>  $ i          : NULL
#>  $ j          : num 1
#>  $ dots       : list()
#>  $ transformed: logi FALSE

# This would be equivalent to calling `x[]`
slicer(x, )

#> List of 4
#>  $ i          : NULL
#>  $ j          : NULL
#>  $ dots       : list()
#>  $ transformed: logi FALSE

# Extra `...` are standardized too. Again, implicit dimensions are returned
# as `NULL`.
slicer(x, 1, 2, 3)

#> List of 4
#>  $ i          : num 1
#>  $ j          : num 2
#>  $ dots       :List of 1
#>   ..$ : num 3
#>  $ transformed: logi FALSE

slicer(x, 1, 2, , 3)

#> List of 4
#>  $ i          : num 1
#>  $ j          : num 2
#>  $ dots       :List of 2
#>   ..$ : NULL
#>   ..$ : num 3
#>  $ transformed: logi FALSE

# You don't have to supply `...` if you don't need it. This
# won't be as useful because it isn't as complex to count here, but it
# will still standardize `x[i]`, which is nice.
slicer_simple <- function(x, i, j) {
  collect_subscripts(i, j)
}

slicer_simple(x, 1)

#> $i
#> NULL
#> 
#> $j
#> [1] 1
#> 
#> $dots
#> list()
#> 
#> $transformed
#> [1] TRUE
#>