Standardize subscripts

collect_subscripts() standardizes subscripts to make it easier to develop [ methods, especially for data frame and array subsetting. It helps with a number of things, including:

Counting the number of subscript arguments supplied.

Transforming implicit missing subscripts into explicit NULL values.

Optional support for capturing ... and drop.

Optional support for standardizing the data frame column subsetting behavior of x[i] into x[,j].

collect_subscripts() can only be called from inside of a function. Calling collect_subscripts() in the console directly will result in an error.

The expected way to use collect_subscripts() is as the first line of a [ method. You'll collect the processed subscripts, and then use them to slice the input accordingly.

collect_subscripts(i, j, ..., drop = NULL, column_transform = TRUE)

Arguments

i, j
[vector]

Required vectors of subscripts.

...
[dots]

Optional additional subscripts for subsetting into higher dimensions. Names are preserved. Implicit missing subscripts are transformed into explicit NULL values.

If no ... are provided, the return value will contain an empty list in the $dots slot.

drop
[NULL / logical(1)]

Optional argument to capture drop. If your [ method signature doesn't include a drop argument, don't supply this argument. If it does, it is required that drop has a default value in your method. It is recommended to default this to drop = FALSE in your method.

column_transform
[logical(1)]

Should a single i argument represent column subsetting? For example, this standardizes x[i] to x[,j]. This is useful when creating data frame [ methods, but array subsetting likely won't use this feature.

Value

A named list of the following 6 elements containing information about the standardized subscripts.

i: The standardized i subscript. If NULL, i is considered to be missing.

j: The standardized j subscript. If NULL, j is considered to be missing.

dots: A list of the standardized ... subscripts. If no ... are supplied, an empty list is returned. If ... are supplied and any subscripts are missing, NULL values are returned in those locations. For example, x[i, j, , k] would return a list of list(NULL, k).

drop: The standardized drop argument. If drop is not provided, this is NULL.

n_args: The total number of subscript-related arguments supplied. This includes all arguments in the [ call except for the object being subset, x. Missing arguments are counted here. For example, x[1,] has a n_args value of 2 because i is supplied and j is supplied but missing. On the other hand, x[i] has a n_args value of 1.

n_subscripts: The total number of subscript arguments supplied. This counts i, j, and ... arguments. It is only different drop n_args if drop is directly supplied by the end user.

Details

When counting arguments, collect_subscripts() assumes that 1 argument is reserved for x and will subtract it from the total number of arguments. If you are implementing a [ method, this is automatically enforced by the method signature. If you are calling collect_subscripts() from somewhere else, keep this in mind when inspecting the results.

If NULL is provided as a subscript, it is standardized to integer() to prevent confusion with NULL values in the resulting info list. A NULL subscript should have the same meaning as integer().

Examples

# Generally you would replace `slicer()` with your `[.myclass` method,
# which you would then call as `x[i]`. This just collects the subscript info
# and returns a compact view of it. Normally you would use the processed
# results to slice `x`.

slicer <- function(x, i, j, ..., drop = FALSE) {
  info <- collect_subscripts(i, j, ..., drop = drop)
  str(info)
}

x <- "a"

# Standardize an easy case
slicer(x, 1:5, 2)

#> List of 6
#>  $ i           : int [1:5] 1 2 3 4 5
#>  $ j           : num 2
#>  $ dots        : list()
#>  $ drop        : logi FALSE
#>  $ n_args      : int 2
#>  $ n_subscripts: int 2

# By default, supplying just `i` standardizes `x[i]` to `x[,j]`
slicer(x, 1)

#> List of 6
#>  $ i           : NULL
#>  $ j           : num 1
#>  $ dots        : list()
#>  $ drop        : logi FALSE
#>  $ n_args      : int 1
#>  $ n_subscripts: int 1

# With just `j`, the implicit missing `i` value is returned as a `NULL`.
# Note that while `i` and `j` are the same for this and `slicer(x, 1)`,
# the `n_*` counters are different.
slicer(x, , 1)

#> List of 6
#>  $ i           : NULL
#>  $ j           : num 1
#>  $ dots        : list()
#>  $ drop        : logi FALSE
#>  $ n_args      : int 2
#>  $ n_subscripts: int 2

# This would be equivalent to calling `x[]`
slicer(x, )

#> List of 6
#>  $ i           : NULL
#>  $ j           : NULL
#>  $ dots        : list()
#>  $ drop        : logi FALSE
#>  $ n_args      : int 1
#>  $ n_subscripts: int 1

# Extra `...` are standardized too. Again, implicit dimensions are returned
# as `NULL`.
slicer(x, 1, 2, 3)

#> List of 6
#>  $ i           : num 1
#>  $ j           : num 2
#>  $ dots        :List of 1
#>   ..$ : num 3
#>  $ drop        : logi FALSE
#>  $ n_args      : int 3
#>  $ n_subscripts: int 3

slicer(x, 1, 2, , 3)

#> List of 6
#>  $ i           : num 1
#>  $ j           : num 2
#>  $ dots        :List of 2
#>   ..$ : NULL
#>   ..$ : num 3
#>  $ drop        : logi FALSE
#>  $ n_args      : int 4
#>  $ n_subscripts: int 4

# You don't have to supply `...` and `drop` if you don't need it. This
# won't be as useful because it isn't as complex to count here, but it
# will still standardize `x[i]`, which is nice.
slicer_simple <- function(x, i, j) {
  collect_subscripts(i, j)
}

# Note that `drop` is returned as `NULL` here, since it isn't in your
# function signature.
slicer_simple(x, 1)

#> $i
#> NULL
#> 
#> $j
#> [1] 1
#> 
#> $dots
#> list()
#> 
#> $drop
#> NULL
#> 
#> $n_args
#> [1] 1
#> 
#> $n_subscripts
#> [1] 1
#>